+++ test/parse/types.h
@@ -1,4 +1,5 @@
 #include <stdbool.h>
+#include <stdint.h>
 
 #include "../../src/parse/types.h"
 #include "../../test/testing.h"
@@ -38,7 +39,7 @@ bool test_make_new_type_rejects_non_unique_type() {
 	const bool inserted1=make_new_type(L"test_type", 1);
 	const bool inserted2=make_new_type(L"test_type", 1);
 
-	unsigned int count=0;
+	uint8_t count=0;
 
 	while (current) {
 		count+=(wcscmp(current->name, L"test_type")==0);
@@ -75,7 +76,7 @@ bool test_append_default_types() {
 	make_default_types();
 	type_t *head=&TYPES_AVAILABLE;
 
-	unsigned int count=0;
+	uint8_t count=0;
 	while (head) {
 		count++;
 		head=head->next;
@@ -84,7 +85,7 @@ bool test_append_default_types() {
 	make_default_types();
 	head=&TYPES_AVAILABLE;
 
-	unsigned int new_count=0;
+	uint8_t new_count=0;
 	while (head) {
 		new_count++;
 		head=head->next;
